+++ container/common/loader.py
@@ -1,18 +1,26 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
-import logging
 import importlib
 
 from container.common.visibility import getLogger
 logger = getLogger(__name__)
 
+from .utils import CAPABILITIES
 
-def load_engine(capabilities_needed, engine_name, project_name, services=[], **kwargs):
+# If "conductor" is a thing, we're inside of a container. If not, we're not.
+try:
+    import conductor
+except ImportError:
+    package = 'container'
+else:
+    package = 'conductor'
+
+def load_engine(capabilities_needed, engine_name, project_name,
+                services=[], **kwargs):
     logger.debug(u"Loading engine capabilities", capabilities=capabilities_needed, engine=engine_name)
-    conductor_module_name = __name__.rsplit('.', 1)[0]
     mod = importlib.import_module('.%s.engine' % engine_name,
-                                  package=conductor_module_name)
+                                  package=package)
     engine_obj = mod.Engine(project_name, services, **kwargs)
     for capability in capabilities_needed:
         if not getattr(engine_obj, 'CAP_%s' % capability):
