+++ container/docker/files/get-pip.py
