+++ container/lib/conductor/docker/engine.py
@@ -9,13 +9,9 @@ logger = getLogger(__name__)
 
 import base64
 import datetime
-import functools
-import inspect
 import json
 import os
 import re
-import six
-import sys
 import tarfile
 
 try:
@@ -25,8 +21,9 @@ except ImportError:
 
 from ..engine import BaseEngine
 from .. import utils
-from .. import logmux
-from .. import exceptions
+from container.common import exceptions, logmux
+from container.common.docker import DockerEngineUtilityMixin
+from container.common.utils import log_runs
 
 try:
     import docker
@@ -34,18 +31,6 @@ try:
     from docker.utils.ports import build_port_bindings
 except ImportError:
     raise ImportError('Use of this engine requires you "pip install \'docker>=2.1\'" first.')
-
-TEMPLATES_PATH = os.path.normpath(
-    os.path.join(
-        os.path.dirname(__file__),
-        'templates'))
-
-FILES_PATH = os.path.normpath(
-    os.path.join(
-        os.path.dirname(__file__),
-        'files'))
-
-DOCKER_VERSION = '1.13.1'
 
 DOCKER_DEFAULT_CONFIG_PATH = os.path.join(os.environ.get('HOME', ''), '.docker', 'config.json')
 
@@ -57,52 +42,8 @@ DOCKER_CONFIG_FILEPATH_CASCADE = [
 
 REMOVE_HTTP = re.compile('^https?://')
 
-def log_runs(fn):
-    @functools.wraps(fn)
-    def __wrapped__(self, *args, **kwargs):
-        logger.debug(
-            u'Call: %s.%s' % (type(self).__name__, fn.__name__),
-            # because log_runs is a decorator, we need to override the caller
-            # line & function
-            caller_func='%s.%s' % (type(self).__name__, fn.__name__),
-            caller_line=inspect.getsourcelines(fn)[-1],
-            args=args,
-            kwargs=kwargs,
-        )
-        return fn(self, *args, **kwargs)
-    return __wrapped__
 
-
-class Engine(BaseEngine):
-
-    # Capabilities of engine implementations
-    CAP_BUILD_CONDUCTOR = True
-    CAP_BUILD = True
-    CAP_DEPLOY = True
-    CAP_IMPORT = True
-    CAP_LOGIN = True
-    CAP_PUSH = True
-    CAP_RUN = True
-
-    display_name = u'Docker\u2122 daemon'
-
-    _client = None
-    _api_client = None
-
-    FINGERPRINT_LABEL_KEY = 'com.ansible.container.fingerprint'
-    LAYER_COMMENT = 'Built with Ansible Container (https://github.com/ansible/ansible-container)'
-
-    @property
-    def client(self):
-        if not self._client:
-            self._client = docker.from_env()
-        return self._client
-
-    @property
-    def api_client(self):
-        if not self._api_client:
-            self._api_client = docker.APIClient()
-        return self._api_client
+class Engine(BaseEngine, DockerEngineUtilityMixin):
 
     @property
     def ansible_args(self):
@@ -125,12 +66,6 @@ class Engine(BaseEngine):
                 result = os.path.normpath(os.path.expanduser(path))
                 break
         return result
-
-    def container_name_for_service(self, service_name):
-        return u'%s_%s' % (self.project_name, service_name)
-
-    def image_name_for_service(self, service_name):
-        return u'%s-%s' % (self.project_name, service_name)
 
     def run_kwargs_for_service(self, service_name):
         to_return = self.services[service_name].copy()
@@ -163,190 +98,6 @@ class Engine(BaseEngine):
         mux = logmux.LogMultiplexer()
         mux.add_iterator(log_iter, plainLogger)
         return container_obj.id
-
-    @log_runs
-    def run_conductor(self, command, config, base_path, params):
-        image_id = self.get_latest_image_id_for_service('conductor')
-        if image_id is None:
-            raise exceptions.AnsibleContainerConductorException(
-                    u"Conductor container can't be found. Run "
-                    u"`ansible-container build` first")
-        serialized_params = base64.encodestring(json.dumps(params))
-        serialized_config = base64.encodestring(json.dumps(config))
-        volumes = {base_path: {'bind': '/src', 'mode': 'ro'}}
-        environ = {}
-        if os.environ.get('DOCKER_HOST'):
-            environ['DOCKER_HOST'] = os.environ['DOCKER_HOST']
-            if os.environ.get('DOCKER_CERT_PATH'):
-                environ['DOCKER_CERT_PATH'] = '/etc/docker'
-                volumes[os.environ['DOCKER_CERT_PATH']] = {'bind': '/etc/docker',
-                                                           'mode': 'ro'}
-            if os.environ.get('DOCKER_TLS_VERIFY'):
-                environ['DOCKER_TLS_VERIFY'] = os.environ['DOCKER_TLS_VERIFY']
-        else:
-            environ['DOCKER_HOST'] = 'unix:///var/run/docker.sock'
-            volumes['/var/run/docker.sock'] = {'bind': '/var/run/docker.sock',
-                                               'mode': 'rw'}
-
-        environ['ANSIBLE_ROLES_PATH'] = '/src/roles:/etc/ansible/roles'
-
-        if params.get('devel'):
-            from container import conductor
-            conductor_path = os.path.dirname(conductor.__file__)
-            logger.debug(u"Binding conductor at %s into conductor container", conductor_path)
-            volumes[conductor_path] = {'bind': '/_ansible/conductor/conductor', 'mode': 'rw'}
-
-        if command in ('login', 'push') and params.get('config_path'):
-            config_path = params.get('config_path')
-            volumes[config_path] = {'bind': config_path,
-                                    'mode': 'rw'}
-
-        run_kwargs = dict(
-            name=self.container_name_for_service('conductor'),
-            command=['conductor',
-                     command,
-                     '--project-name', self.project_name,
-                     '--engine', __name__.rsplit('.', 2)[-2],
-                     '--params', serialized_params,
-                     '--config', serialized_config,
-                     '--encoding', 'b64json'],
-            detach=True,
-            user='root',
-            volumes=volumes,
-            environment=environ,
-            working_dir='/src',
-            cap_add=['SYS_ADMIN']
-        )
-
-        logger.debug('Docker run:', image=image_id, params=run_kwargs)
-
-        try:
-            container_obj = self.client.containers.run(
-                image_id,
-                **run_kwargs
-            )
-        except docker_errors.APIError as exc:
-            if exc.response.status_code == StatusCodes.CONFLICT:
-               raise exceptions.AnsibleContainerConductorException(
-                    u"Can't start conductor container, another conductor for "
-                    u"this project already exists or wasn't cleaned up.")
-            six.reraise(*sys.exc_info())
-        else:
-            log_iter = container_obj.logs(stdout=True, stderr=True, stream=True)
-            mux = logmux.LogMultiplexer()
-            mux.add_iterator(log_iter, plainLogger)
-            return container_obj.id
-
-    def service_is_running(self, service):
-        try:
-            container = self.client.containers.get(self.container_name_for_service(service))
-            return container.status == 'running' and container.id
-        except docker_errors.NotFound:
-            return False
-
-    def service_exit_code(self, service):
-        try:
-            container = self.client.api.inspect_container(self.container_name_for_service(service))
-            return container['State']['ExitCode']
-        except docker_errors.APIError:
-            return None
-
-    def stop_container(self, container_id, forcefully=False):
-        try:
-            container = self.client.containers.get(container_id)
-        except docker_errors.APIError:
-            pass
-        else:
-            if forcefully:
-                container.kill()
-            else:
-                container.stop(timeout=60)
-
-    def restart_all_containers(self):
-        raise NotImplementedError()
-
-    def inspect_container(self, container_id):
-        try:
-            return self.client.api.inspect_container(container_id)
-        except docker_errors.APIError:
-            return None
-
-    def delete_container(self, container_id):
-        try:
-            container = self.client.containers.get(container_id)
-        except docker_errors.APIError:
-            pass
-        else:
-            container.remove()
-
-    def get_container_id_for_service(self, service_name):
-        try:
-            container = self.client.containers.get(self.container_name_for_service(service_name))
-        except docker_errors.NotFound:
-            return None
-        else:
-            return container.id
-
-    def get_image_id_by_fingerprint(self, fingerprint):
-        try:
-            image, = self.client.images.list(
-                all=True,
-                filters=dict(label='%s=%s' % (self.FINGERPRINT_LABEL_KEY,
-                                              fingerprint)))
-        except ValueError:
-            return None
-        else:
-            return image.id
-
-    def get_image_id_by_tag(self, tag):
-        try:
-            image = self.client.images.get(tag)
-            return image.id
-        except docker_errors.ImageNotFound:
-            return None
-
-    def get_latest_image_id_for_service(self, service_name):
-        image = self.get_latest_image_for_service(service_name)
-        if image is not None:
-            return image.id
-        return None
-
-    def get_latest_image_for_service(self, service_name):
-        try:
-            image = self.client.images.get(
-                '%s:latest' % self.image_name_for_service(service_name))
-        except docker_errors.ImageNotFound:
-            images = self.client.images.list(name=self.image_name_for_service(service_name))
-            logger.debug("Could not find the latest image for service, "
-                "searching for other tags with same image name",
-                image_name=self.image_name_for_service(service_name),
-                service=service_name)
-
-            if not images:
-                return None
-
-            def tag_sort(i):
-                return [t for t in i.tags if t.startswith(self.image_name_for_service(service_name))][0]
-
-            images = sorted(images, key=tag_sort)
-            logger.debug('Found images for service',
-                    service=service_name,
-                    images=images)
-            return images[-1]
-        else:
-            return image
-
-    def get_build_stamp_for_image(self, image_id):
-        build_stamp = None
-        try:
-            image = self.client.images.get(image_id)
-        except docker_errors.ImageNotFound:
-            raise exceptions.AnsibleContainerConductorException(
-                "Unable to find image {}".format(image_id)
-            )
-        if image and image.tags:
-            build_stamp = [tag for tag in image.tags if not tag.endswith(':latest')][0].split(':')[-1]
-        return build_stamp
 
     @log_runs
     def commit_role_as_layer(self,
@@ -443,88 +194,6 @@ class Engine(BaseEngine):
                     last_status = line['status']
                 else:
                     plainLogger.debug(line)
-
-    @log_runs
-    def build_conductor_image(self, base_path, base_image, cache=True):
-        with utils.make_temp_dir() as temp_dir:
-            logger.info('Building Docker Engine context...')
-            tarball_path = os.path.join(temp_dir, 'context.tar')
-            tarball_file = open(tarball_path, 'wb')
-            tarball = tarfile.TarFile(fileobj=tarball_file,
-                                      mode='w')
-            source_dir = os.path.normpath(base_path)
-
-            for filename in ['ansible.cfg', 'ansible-requirements.txt',
-                             'requirements.yml']:
-                file_path = os.path.join(source_dir, filename)
-                if os.path.exists(filename):
-                    tarball.add(file_path,
-                                arcname=os.path.join('build-src', filename))
-            # Make an empty file just to make sure the build-src dir has something
-            open(os.path.join(temp_dir, '.touch'), 'w')
-            tarball.add(os.path.join(temp_dir, '.touch'), arcname='build-src/.touch')
-
-            tarball.add(os.path.join(FILES_PATH, 'get-pip.py'),
-                        arcname='contrib/get-pip.py')
-
-            tarball.add(utils.conductor_dir, arcname='conductor-src/conductor')
-            tarball.add(os.path.join(os.path.dirname(utils.conductor_dir),
-                                     'conductor-setup.py'),
-                        arcname='conductor-src/setup.py')
-            tarball.add(os.path.join(os.path.dirname(utils.conductor_dir),
-                                     'conductor-requirements.txt'),
-                        arcname='conductor-src/requirements.txt')
-
-            utils.jinja_render_to_temp(TEMPLATES_PATH,
-                                       'conductor-dockerfile.j2', temp_dir,
-                                       'Dockerfile',
-                                       conductor_base=base_image,
-                                       docker_version=DOCKER_VERSION)
-            tarball.add(os.path.join(temp_dir, 'Dockerfile'),
-                        arcname='Dockerfile')
-
-            #for context_file in ['builder.sh', 'ansible-container-inventory.py',
-            #                     'ansible.cfg', 'wait_on_host.py', 'ac_galaxy.py']:
-            #    tarball.add(os.path.join(TEMPLATES_PATH, context_file),
-            #                arcname=context_file)
-
-            logger.debug('Context manifest:')
-            for tarinfo_obj in tarball.getmembers():
-                logger.debug('tarball item: %s (%s bytes)', tarinfo_obj.name,
-                             tarinfo_obj.size, file=tarinfo_obj.name,
-                             bytes=tarinfo_obj.size, terse=True)
-            tarball.close()
-            tarball_file.close()
-            tarball_file = open(tarball_path, 'rb')
-            logger.info('Starting Docker build of Ansible Container Conductor image (please be patient)...')
-            # FIXME: Error out properly if build of conductor fails.
-            if self.debug:
-                for line in self.client.api.build(fileobj=tarball_file,
-                                                  custom_context=True,
-                                                  tag=self.image_name_for_service('conductor'),
-                                                  rm=True,
-                                                  nocache=not cache):
-                    try:
-                        line_json = json.loads(line)
-                        if 'stream' in line_json:
-                            line = line_json['stream']
-                        elif line_json.get('status') == 'Downloading':
-                            # skip over lines that give spammy byte-by-byte
-                            # progress of downloads
-                            continue
-                    except ValueError:
-                        pass
-                    # this bypasses the fancy colorized logger for things that
-                    # are just STDOUT of a process
-                    plainLogger.debug(line.rstrip())
-                return self.get_latest_image_id_for_service('conductor')
-            else:
-                image = self.client.images.build(fileobj=tarball_file,
-                                                 custom_context=True,
-                                                 tag=self.image_name_for_service('conductor'),
-                                                 rm=True,
-                                                 nocache=not cache)
-                return image.id
 
     def get_runtime_volume_id(self):
         try:
