+++ container/shipit/kubernetes/engine.py
@@ -2,15 +2,15 @@
 
 from __future__ import absolute_import
 
+import json
 import logging
 import os.path
-import json
 
-from container.utils import create_path
-from ..base_engine import BaseShipItEngine
+from container.common.utils import create_path
 from .deployment import Deployment
-from .service import Service
 from .pvc import Pvc
+from .service import Service
+from ..base_engine import BaseShipItEngine
 from ..constants import SHIPIT_PATH, SHIPIT_CONFIG_PATH
 
 logger = logging.getLogger(__name__)
